requirements.txt
absl-py==2.3.1
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
branca==0.8.2
cachetools==6.2.2
cdsapi==0.7.7
certifi==2025.11.12
cffi==2.0.0
cfgrib==0.9.15.1
cftime==1.6.5
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
eccodes==2.44.0
ecmwf-datastores-client==0.4.1
fastapi==0.124.4
filelock==3.20.0
findlibs==0.1.2
flatbuffers==25.9.23
folium==0.20.0
fonttools==4.61.0
fsspec==2025.12.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5netcdf==1.7.3
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
multiurl==0.3.7
namex==0.1.0
narwhals==2.13.0
netCDF4==1.7.3
networkx==3.6.1
numpy==2.3.5
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pluggy==1.6.0
protobuf==6.33.2
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
starlette==0.50.0
streamlit==1.52.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
uvicorn==0.38.0
watchdog==6.0.0
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
xarray==2025.12.0
xgboost==3.1.3
xyzservices==2025.11.0
